Change Custody Count to Uint8 (#14386)
* Add Changes for Uint8 Csc

* Fix Build

* Fix Build for Sync

* Fix Discovery Test

+++ proto/prysm/v1alpha1/metadata/metadata_interfaces.go
@@ -11,7 +11,7 @@ type Metadata interface {
 	SequenceNumber() uint64
 	AttnetsBitfield() bitfield.Bitvector64
 	SyncnetsBitfield() bitfield.Bitvector4
-	CustodySubnetCount() uint64
+	CustodySubnetCount() uint8
 	InnerObject() interface{}
 	IsNil() bool
 	Copy() Metadata
